requirements.txt
streamlit==1.37.0
requests==2.32.3
python-dotenv==1.0.1
altair==5.4.1 
langchain-groq==0.1.9
scikit-learn==1.6.0